Extract Shell and Batch reference helpers
